Op.15 No.5: proofreading
Measures 18 and 19, left hand: In the middle voice of the left hand, the
note tied from measure 18 to 19 should be G-natural instead of G-sharp.
Compare these two measures to measures 10 and 11, which should be
identical.

Measure 20, left hand: The final note of middle voice should be
F-natural instead of F-sharp. Compare this measure to measure 12, which
should be identical.